Skip nondeterministic moves
This solves all of the basic puzzles except Chandelier and some of the
helium puzzles.  Chandelier seems to require a nondeterministic move, but
the puzzle's symmetry makes the nondeterminism irrelevant.  The helium
puzzles are symmetric and our naive search will double or quadruple its
effort searching equivalent branches of the state tree; we need to
normalize states for symmetries to make progress there.