Unboxed representation
This is much faster (3x+).  We're using a less-than-canonical
representation for packed states (we don't sort), but we still manage to
get to generation 36 because our packed rep is a little denser (once we
have a concept of an atom index we can store just the coordinate per atom
and a bond is just an index pair).