Forbidden bonds pruning
We now solve Three Doors (45 generations, 54471 ms, 22944451 closed set
states) and Planning (26 generations, 78445 ms, 2091635 states in the
closed set).  (Times/states are for the parallel runs, so are somewhat
noisy.)

We solve all puzzles not using bonders/splitters except for Chandelier,
which is "nondeterministic": there are two possible bonds but only one
free electron, and we don't know which bond we will form.  (Sokobond
consistently bonds to the right (at least for this one puzzle), so the
obvious choice, row-major order, is not correct.)